added RTC code
This isn't quite working yet, but I thought I'd upload it anyway just so we have it. Essentially it reads and sets the time, but it just seems to run a loop in which nothing changes at the moment so I'll work on it. If you're going to be trying to change any code, go for the display, probably, as that's definitely working and is going to need the most work to get it doing what we want.