added heartbeat sensor code
This ends up producing quite a wide range of outputs in a very short space of time, so I think it would probably be a good idea to take a mean of several readings before outputting a value. Also, this probably only needs to be running while you're checking for a pulse, because otherwise there will be a red LED perpetually on that is likely to be both irritating and a power drain.