added code for Bluetooth connection
This requires a specific app called Bluefruit - Adafruit's Bluetooth setup, but allows you to send any text data you like as if you were typing it into the console, I believe. This code currently only spits it back out in the console on the computer, but if we wanted to we could use it to take in a string for time and apply it to the RTC, perhaps even allow it to take the time directly from the connected device.