replaced battery code
I started from scratch and rewrote this - the calculation used to convert the analogue read value is based on the numbers I needed to accurately record the voltage produced by the battery I was using, this may need changing slightly depending on what battery we end up using. Another thing to note is that lithium batteries, which we will be using, don't have much in the way of a noticeable voltage drop before they die - about 0.1V, usually,  so having maybe 3 bars of battery level will probably better reflect the level of accuracy we can work with than a percentage.