added code to switch off display backlight
it's not much different from the ordinary test code because that's what I based it on, but the important things to note are the assignment of backlight_pin, which is switched to low when you want to turn the backlight off. I haven't checked that the screen still runs okay after turning the backlight on and off again but that's worth looking into.